+++ src/main/java/com/sscanner/team/auth/service/CustomUserDetailsService.kt
@@ -3,7 +3,6 @@ package com.sscanner.team.auth.service
 import com.sscanner.team.global.exception.BadRequestException
 import com.sscanner.team.global.exception.ExceptionCode
 import com.sscanner.team.user.repository.UserRepository
-import com.sscanner.team.user.responsedto.UserDetailsImpl
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.security.core.userdetails.UserDetailsService
 import org.springframework.security.core.userdetails.UsernameNotFoundException
