+++ src/main/java/com/sscanner/team/global/common/BaseEntity.java
@@ -1,28 +1,28 @@
-package com.sscanner.team.global.common
+package com.sscanner.team.global.common;
 
-import jakarta.persistence.Column
-import jakarta.persistence.EntityListeners
-import jakarta.persistence.MappedSuperclass
-import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.LastModifiedDate
-import org.springframework.data.jpa.domain.support.AuditingEntityListener
-import java.time.LocalDateTime
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-@EntityListeners(AuditingEntityListener::class)
+import java.time.LocalDateTime;
+
+@EntityListeners(AuditingEntityListener.class)
 @MappedSuperclass
-open class BaseEntity {
+@Getter
+public class BaseEntity {
 
     @CreatedDate
-    @Column(name = "created_at", updatable = false)
-    var createdAt: LocalDateTime? = null
-        private set
+    @Column(name = "created_at")
+    private LocalDateTime createdAt;
 
     @LastModifiedDate
     @Column(name = "updated_at")
-    var updatedAt: LocalDateTime? = null
-        private set
+    private LocalDateTime updatedAt;
 
     @Column(name = "deleted_at")
-    var deletedAt: LocalDateTime? = null
-        private set
+    private LocalDateTime deletedAt;
 }
